requirements.txt
altgraph==0.17.4
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2025.1.31
click==8.1.8
Eel==0.18.1
flaskwebgui==1.1.6
flet==0.27.6
fpdf==1.7.2
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
htmldocx==0.0.6
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.2
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
pillow==11.2.1
psutil==6.1.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.1
python-docx==1.1.2
repath==0.9.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
